Import libptp++
For various reasons, I've decided to rename libptp++ to libEasyPTP.
This is basically just a carry-over of the libptp++ repository (the backend-
interfaces branch specifically), with a few minor name changes included.

Coming up, I'll be making the rest of the renames, as well as converting
this to a Makefile project and moving some files around.